revert reliance on uuidcomponent for layers
