+++ packages/ecs/src/UUIDComponent.ts
@@ -23,6 +23,86 @@ All portions of the code written by the Infinite Reality Engine team are Copyrig
 Infinite Reality Engine. All Rights Reserved.
 */
 
-import { UUIDComponent } from './ComponentFunctions'
+import { NO_PROXY_STEALTH, State, hookstate, useHookstate } from '@ir-engine/hyperflux'
+import { v4 as uuidv4 } from 'uuid'
+import { LayerComponent, LayerID, Layers, defineComponent, setComponent } from './ComponentFunctions'
+import { Entity, EntityUUID, UndefinedEntity } from './Entity'
+import { createEntity } from './createEntity'
+import { S } from './schemas/JSONSchemas'
 
-export { UUIDComponent }
+export const UUIDComponent = defineComponent({
+  name: 'UUIDComponent',
+  jsonID: 'EE_uuid',
+
+  schema: S.Required(
+    S.EntityUUID({
+      validate: (uuid, prev, entity) => {
+        if (!uuid) {
+          console.error('UUID cannot be empty')
+          return false
+        }
+        if (uuid === prev) return true
+        const layer = LayerComponent.get(entity)
+        // throw error if uuid is already in use
+        const currentEntity = UUIDComponent.getEntityByUUID(uuid, layer)
+        if (currentEntity !== UndefinedEntity && currentEntity !== entity) {
+          console.error(`UUID ${uuid} is already in use`)
+          return false
+        }
+
+        // remove old uuid
+        if (prev) {
+          const currentUUID = prev
+          _getUUIDState(currentUUID, layer).set(UndefinedEntity)
+        }
+
+        // set new uuid
+        _getUUIDState(uuid, layer).set(entity)
+        return true
+      }
+    })
+  ),
+
+  onRemove: (entity, component) => {
+    const uuid = component.value
+    const layer = LayerComponent.get(entity)
+    _getUUIDState(uuid, layer).set(UndefinedEntity)
+  },
+
+  entitiesByUUIDState: {} as Record<LayerID, Record<EntityUUID, State<Entity>>>,
+
+  useEntityByUUID(uuid: EntityUUID, layer = Layers.Simulation as LayerID) {
+    return useHookstate(_getUUIDState(uuid, layer)).value
+  },
+
+  getEntityByUUID(uuid: EntityUUID, layer = Layers.Simulation as LayerID) {
+    return _getUUIDState(uuid, layer).get(NO_PROXY_STEALTH)
+  },
+
+  getOrCreateEntityByUUID(uuid: EntityUUID, layer = Layers.Simulation as LayerID) {
+    const state = _getUUIDState(uuid, layer)
+    if (!state.value) {
+      const entity = createEntity(layer)
+      setComponent(entity, UUIDComponent, uuid)
+    }
+    return state.value
+  },
+
+  generateUUID() {
+    return uuidv4() as EntityUUID
+  }
+})
+
+function _getUUIDState(uuid: EntityUUID, layer = Layers.Simulation as LayerID) {
+  let layerState = UUIDComponent.entitiesByUUIDState[layer]
+  if (!layerState) {
+    layerState = {}
+    UUIDComponent.entitiesByUUIDState[layer] = layerState
+  }
+  let entityState = layerState[uuid]
+  if (!entityState) {
+    entityState = hookstate(UndefinedEntity)
+    layerState[uuid] = entityState
+  }
+  return entityState
+}
